further split the project into more modules
