+++ src/canisters/frontend/ic-asset/src/batch_upload/plumbing.rs
@@ -1,9 +1,6 @@
-use crate::canister_api::{
-    protocol::AssetDetails,
-    chunk::create_chunk
-};
 use crate::asset::{config::AssetConfig, content::Content, content_encoder::ContentEncoder};
 use crate::batch_upload::semaphores::Semaphores;
+use crate::canister_api::{methods::chunk::create_chunk, types::asset::AssetDetails};
 use candid::Nat;
 use futures::future::try_join_all;
 use futures::TryFutureExt;
