--- src/canisters/frontend/ic-asset/src/canister_api/method_names.rs
@@ -1,5 +1,0 @@
-pub(crate) const API_VERSION: &str = "api_version";
-pub(crate) const CREATE_BATCH: &str = "create_batch";
-pub(crate) const CREATE_CHUNK: &str = "create_chunk";
-pub(crate) const COMMIT_BATCH: &str = "commit_batch";
-pub(crate) const LIST: &str = "list";
