+++ src/canisters/frontend/ic-asset/src/canister_api/methods/api_version.rs
@@ -1,4 +1,4 @@
-use crate::canister_api::method_names::API_VERSION;
+use crate::canister_api::methods::method_names::API_VERSION;
 
 use ic_utils::call::SyncCall;
 use ic_utils::Canister;
