+++ src/canisters/frontend/ic-asset/src/canister_api/methods/chunk.rs
@@ -2,8 +2,8 @@ use std::time::Duration;
 
 use crate::batch_upload::{retryable::retryable, semaphores::Semaphores};
 use crate::canister_api::{
-    method_names::CREATE_CHUNK,
-    protocol::{CreateChunkRequest, CreateChunkResponse},
+    methods::method_names::CREATE_CHUNK,
+    types::batch_upload::common::{CreateChunkRequest, CreateChunkResponse},
 };
 use anyhow::bail;
 use backoff::backoff::Backoff;
