+++ src/canisters/frontend/ic-asset/src/canister_api/mod.rs
@@ -1,6 +1,2 @@
-pub(crate) mod api_version;
-pub(crate) mod batch;
-pub(crate) mod chunk;
-pub(crate) mod list;
-pub(crate) mod method_names;
-pub(crate) mod protocol;
+pub(crate) mod types;
+pub(crate) mod methods;
