+++ src/canisters/frontend/ic-asset/src/canister_api/types/batch_upload/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod common;
+pub(crate) mod v0;
