+++ src/canisters/frontend/ic-asset/src/canister_api/types/list.rs
@@ -1,0 +1,5 @@
+use candid::CandidType;
+
+/// Return a list of all assets in the canister.
+#[derive(CandidType, Debug)]
+pub struct ListAssetsRequest {}
