+++ src/canisters/frontend/ic-asset/src/canister_api/types/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod asset;
+pub(crate) mod batch_upload;
+pub(crate) mod list;
